+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### 0.10 (WIP)
 * reworked `*.iml`, `*.war` files.
-* support `*.toml`, `*.asciidoc`, `*.versionsbackup` files.
+* support `nginx.conf`, `.htaccess`, `puppet.conf`, `my.ini`, `*.ini`, `*.cfg`, `*.conf`, `*.toml`, `*.asciidoc`, `*.versionsbackup` files.
 
 ### 0.9 (2018/09/15)
 * reworked `LICENSE`, `berksfile.lock` files.